fix: ColorPickerFrame game version compatibility
This addons calls `SetColor` in every swatchFunc call.

`SetColor` updates the swatch texture color, and it also saves the color to the db **only if the ColorPickerFrame is closed**.

Clicking "Okay" button invokes the swatchFunc callback on both retail and classic

On classic the color picker is hidden before the swatchFunc is called. Howeve, on retail the swatchFunc is called **before** the color picker is hidden.

As a result, in retail the color from the picker is not being saved because the ColorPickerFrame was still open when the swatchFunc was called and invoked `SetColor`.