major: allow for pre-exisitng `CUSTOM_CLASS_COLORS`
Since this addon was not supported for a while, there have been other addons that offer support for forcing colors for certain classes (primarily thinking shaman in classic).

These addons create a `CUSTOM_CLASS_COLORS` table as well and I think it would be wise allow for an existing table.

defaults should still be the blizzard defaults.